ボール、的、跳ね返り処理作成 的の挙動更新 不要ファイル削除

+++ Assets/batting/Components/TargetInfo.cs
@@ -9,6 +9,6 @@ namespace Batting
 		public int Status;
 		public float Timer;
 		public float Radius;
-		public float Level;
+		public int Level;
 	}
 }
